WIP: S29 Plugin Management CLI - cycle 1, CLI commands complete
- Add prb plugins subcommand with list/info/install/remove
- Implement plugin directory management (~/.prb/plugins)
- Add --plugin-dir and --no-plugins global flags
- Update CLI tests to use --where-clause instead of --where
- Fix clippy warnings in prb-capture (is_multiple_of, filter_map)
- All 29 CLI tests passing

Note: Runtime plugin loading in prb ingest (T6.6) deferred to future work.
CLI infrastructure complete, plugin management commands functional.

+++ crates/prb-pcap/src/parallel/normalize.rs
@@ -291,7 +291,7 @@ mod tests {
         let sequential_result: Vec<OwnedNormalizedPacket> = packets
             .iter()
             .filter_map(|pkt| {
-                match normalize_stateless(pkt.linktype as u32, pkt.timestamp_us, &pkt.data) {
+                match normalize_stateless(pkt.linktype, pkt.timestamp_us, &pkt.data) {
                     Ok(NormalizeResult::Packet(p)) => Some(p),
                     _ => None,
                 }
